Move backoff file from /usr/share to /var/run
In `hier(7)` `/usr` is described as:

This directory is usually mounted from a separate partition. It should
hold only shareable, read-only data, so that it can be mounted by
various machines running Linux.

The read-only part is very noticeable on systems like Fedora Silverblue
where only `/var` is writeable and the rest is read-only[0].

`/var/run` is a writable location for data relevant to the current
system boot (its content is cleared on boot).

[0] https://docs.fedoraproject.org/en-US/fedora-silverblue/technical-information/#filesystem-layout